Fixing import paths based on renamed test collection in various Python SDK libraries.
