+++ app/tests/sdk_python/init/test_create.py
@@ -1,4 +1,4 @@
-from tests.sdk.base import BaseTest
+from tests.sdk_python.base import BaseTest
 
 
 class CreateTest(BaseTest):
